refactor(platform-gitlab): 适配器按领域拆分为四个服务
重构 Phase 2c（GitLab）。把单体 GitLabAdapter 拆为四个领域服务（GitLabConnection /
GitLabPullRequestService / GitLabCommentService / GitLabMediaService，各 extends
platform-core 领域基类），共享一份连接上下文；GitLabAdapter 退化为薄门面，过渡期委托
扁平方法并暴露 .connection/.pulls/.comments/.media。

- 连接配置归连接层自管：client 自管 base 归一、clone 协议、web/git host 派生与
  getCloneUrl；edition 探测态 approvalsAvailable 落在 client 上，供连接（capabilities）
  与 PR（审批拉取）领域共读。
- 映射就近维护：领域专属映射下沉为各服务私有方法，公共响应类型抽到 types.ts，
  跨领域 mapUser / projectId 放 utils.ts。

公开方法签名不变、适配器测试全过、desktop typecheck 通过，行为等价。
